pyproject.toml
[tool.poetry]
name = "tailoredscoop"
version = "0.1.0"
description = ""
authors = ["Chansoo Song"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
pytest = "^7.2.0"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
mkdocs = "^1.4.2"
mkdocs-material = ">=8.0.0,<9.0.0"
mkdocs-simple-blog = "^0.0.5"
mkdocstrings = "^0.20.0"
mkdocs-jupyter = "^0.22.0"
pymdown-extensions = "^9.9.2"
langchain = "^0.0.151"
openai = "^0.27.5"
ipykernel = "^6.22.0"
pyyaml = "^6.0"
chromadb = "^0.3.21"
tiktoken = "^0.3.3"
beautifulsoup4 = "^4.12.2"
transformers = "^4.28.1"
datasets = "^2.12.0"
torch = "^2.0.0"
protobuf = "^3.20"
pymongo = "^4.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"